requirements.txt
streamlit>=1.28.0
cartesia>=1.0.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.18.0
